Change review score from Int to Float

+++ src/main/kotlin/io/github/zeroone3010/caffeinegoat/entity/Review.kt
@@ -21,7 +21,7 @@ class Review : PanacheEntity() {
     var reviewer: User = User()
 
     @Column(nullable = false)
-    var rating: Int = 0
+    var rating: Float = 0f
 
     @Column(length = 2048)
     var comment: String? = null
